Make compose_namespace and install_model_shortcut configurable.
This is a long running orphaned topic branch, thrown into mess
by rkitover's switch to defaulting them to off.
Instead, we're making them configurable, without changing the default.
Next step is to deprecate and warn on them, and then change the default.
That can happen over the next year, or not at all if there's no pressure for it.